refactor: wip - proptypes for the form, cleanup
